requirements.txt
python-telegram-bot==20.8
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
dnspython==2.5.0
python-whois==0.8.0
requests==2.31.0
reportlab==4.0.9
beautifulsoup4==4.12.3
email-validator==2.1.0.post1
python-nmap==0.7.1
# ...existing requirements...